requirement.txt
aiofiles==0.7.0
blinker==1.4
Brotli==1.0.9
click==8.0.3
colorama==0.4.4
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
plotly==5.3.1
priority==2.0.0
python-dateutil==2.8.2
pytz==2021.3
Quart==0.15.1
quart-compress==0.2.1
six==1.16.0
tenacity==8.0.1
toml==0.10.2
websocket-client==1.2.1
Werkzeug==2.0.2
wsproto==1.0.0
